pyproject.toml
[tool.poetry]
name = "leetcode"
version = "0.1.0"
description = "Build Leetcode calendar and track your progress"
authors = ["David N <david@davidngugi.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
arrow = "^1.2.3"
requests = "^2.31.0"
selenium = "^4.12.0"
click = "^8.1.7"
google-api-python-client = "^2.100.0"
google-auth-oauthlib = "^1.1.0"
asyncclick = "^8.1.3.4"
anyio = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
